backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlmodel==0.0.14
sqlalchemy[asyncio]==2.0.25
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pydantic[email]==2.5.3
pydantic-settings==2.1.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
websockets==12.0
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.26.0
